requirements.txt
aniso8601==9.0.1
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Deprecated==1.2.13
Flask==2.0.2
Flask-RESTful==0.3.9
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lxml==4.7.1
MarkupSafe==2.0.1
packaging==21.3
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-decouple==3.5
pytz==2021.3
queuelib==1.6.2
redis==4.1.1
requests==2.27.1
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
Werkzeug==2.0.2
wrapt==1.13.3
zope.interface==5.4.0
